+++ src/chronotext/font/zf/FontManager.cpp
@@ -303,9 +303,9 @@ namespace chr
             }
         }
         
-        size_t FontManager::getTextureMemoryUsage() const
+        int64_t FontManager::getTextureMemoryUsage() const
         {
-            size_t total = 0;
+            int64_t total = 0;
             
             for (auto &actualFont : actualFonts)
             {
